# docs/api/openapi.yaml
openapi: 3.0.1
info:
  title: EasyConf Reviews Microservice
  description: "EasyConf is web-based conference management system, designed to organize\
    \ and administrate academic conferences, workshops, and events. It offers a platform\
    \ for authors, reviewers, and organizers (chairs) to handle paper submissions,\
    \ peer reviews, and program scheduling efficiently. It provides tools for managing\
    \ the review process, generating program schedules, and handling communication\
    \ between organizers and participants.\n\nThe Reviews Microservice handles the\
    \ process of reviewing papers before a conference. This microservice is responsible\
    \ for assigning papers to reviews, allowing reviewers to add their comments, and\
    \ allowing the PC chairs to finalize the decision about acceptance and rejection\
    \ of the submitted papers.\n\nThe different roles that the users can have are\
    \ specified in the Users Microservice and the process for submitting papers to\
    \ conferences is specified in the Submissions Microservice.\n\nTECHNICAL NOTE:\
    \ in every endpoint, we require the requesting user to provide their ID. I.e.,\
    \ the requesterID parameter must be provided. It is used for authentication in\
    \ order to check if the user is allowed to perform certain actions."
  version: 0.0.1-SNAPSHOT
servers:
- url: http://localhost:8080
  description: Generated server url
...
paths:
...
  /conferences/{conferenceID}/tracks/{trackID}/automatic:
    put:
      tags:
      - Assignments
      summary: Automatically assign the reviewers
      description: "Automatically assigns a reviewer to a specific paper. At least\
        \ 3 reviewers will be assigned to each paper (taking into account the manual\
        \ assignments as well), such that each reviewer in the track has a similar\
        \ amount of reviews assigned to them."
      operationId: assignAuto
      parameters:
      - name: requesterID
        in: query
        description: The ID of the user making the request
        required: true
        schema:
          type: integer
          format: int64
      - name: conferenceID
        in: path
        description: The ID of the conference the track belongs to
        required: true
        schema:
          type: integer
          format: int64
      - name: trackID
        in: path
        description: The ID of the track for which to do the automatic assignment
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: Reviewers successfully assigned to the track.
        "500":
          description: Internal Server Error. An unexpected server error occurred.
        "409":
          description: Conflict. Automatic assignment not possible. There must be
            at least one reviewer in the track.
        "403":
          description: Forbidden. You are not allowed to assign reviewers. Only the
            chairs for tracks are allowed.
        "404":
          description: Not Found. The specified track or user does not exist.
  /papers/{paperID}/reviews/finalization:
    post:
      tags:
      - Reviews
      summary: End the discussion phase for a paper
      description: "The paper is now considered reviewed: the reviews cannot be changed.\
        \ Once all of the papers in a track are reviewed, the track automatically\
        \ goes to the FINAL phase."
      operationId: finalization
      parameters:
      - name: requesterID
        in: query
        description: The ID of a user making the request
        required: true
        schema:
          type: integer
          format: int64
      - name: paperID
        in: path
        description: The ID of a paper to finalize the reviews for
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "409":
          description: Conflict. The paper must be in its Discussion phase and all
            of the reviews need to either be positive or negative. Only then can they
            be finalized.
        "500":
          description: Internal Server Error. An unexpected server error occurred.
        "200":
          description: Successful finalization of the discussion phase
        "403":
          description: Forbidden. The requester must be a chair of the track the paper
            is in.
        "404":
          description: Not Found. The requested paper was not found.
...
  /papers/{paperID}/assignees/{reviewerID}:
    post:
      tags:
      - Assignments
      summary: Manually assign reviewers
      description: Manually assigns reviewer to a specific paper.At least 3 reviewers
        must be assigned to a paper.
      operationId: assignManual
      parameters:
      - name: requesterID
        in: query
        description: The ID of a user making the request
        required: true
        schema:
          type: integer
          format: int64
      - name: reviewerID
        in: path
        description: The ID of a user to assign as a reviewer
        required: true
        schema:
          type: integer
          format: int64
      - name: paperID
        in: path
        description: The ID of a paper to assign
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "500":
          description: Internal Server Error. An unexpected server error occurred.
        "201":
          description: Reviewer successfully assigned to the paper.
        "403":
          description: Forbidden. You are not allowed to assign reviewers. Only chairs
            for track can do that
        "404":
          description: Not Found. The specified paper or user does not exist.
        "409":
          description: Conflict. There is a Conflict of Interest and the reviewer
            cannot review this paper; or the reviewer has already been assigned to
            this paper.
    delete:
      tags:
      - Assignments
      summary: Removes a reviewer from a paper
      description: Removes a reviewer previously assigned to a paper.
      operationId: remove
      parameters:
      - name: requesterID
        in: query
        description: The ID of a user making the request
        required: true
        schema:
          type: integer
          format: int64
      - name: paperID
        in: path
        description: The ID of a paper reviewer belongs to
        required: true
        schema:
          type: integer
          format: int64
      - name: reviewerID
        in: path
        description: The ID of a reviewer to remove
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: "Not Found. The specified paper or user does not exist, or\
            \ the user is not assigned to this paper."
        "500":
          description: Internal Server Error. An unexpected server error occurred.
        "403":
          description: Forbidden. You are not allowed to remove a reviewer. Only chairs
            can do that.
        "200":
          description: Successfully removed the reviewer for this paper.
  /conferences/{conferenceID}/tracks/{trackID}/finalization:
    post:
      tags:
      - Assignments
      summary: Finalise reviewers
      description: "Finalises the assignment of reviewers, so they can no longer be\
        \ changed manually or automatically. "
      operationId: finalization_1
      parameters:
      - name: requesterID
        in: query
        description: The ID of the user making the request
        required: true
        schema:
          type: integer
          format: int64
      - name: conferenceID
        in: path
        description: The ID of the conference the track belongs to
        required: true
        schema:
          type: integer
          format: int64
      - name: trackID
        in: path
        description: The ID of the track for which the assignments should be finalized
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "201":
          description: Reviewers has been successfully finalized to this paper.
        "500":
          description: Internal Server Error. An unexpected server error occurred.
        "403":
          description: Forbidden. You are not allowed to finalize the assignments.
            You must be a chair of the track.
        "409":
          description: Conflict. There must be at least 3 reviewers assigned to each
            paper to finalize the assignments.
        "404":
          description: Not Found. The specified paper or user does not exist.
...
  /papers/{paperID}/assignees:
    get:
      tags:
      - Assignments
      summary: Get current assignments for a paper
      description: Responds with a list of reviewer IDs for a specific paper.
      operationId: assignments
      parameters:
      - name: requesterID
        in: query
        description: The ID of a user making the request
        required: true
        schema:
          type: integer
          format: int64
      - name: paperID
        in: path
        description: The ID of a paper assignments belong to
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "500":
          description: Internal Server Error. An unexpected server error occurred.
        "200":
          description: Successfully returned the list of reviewers assigned to this
            paper.
        "403":
          description: Forbidden. You are not allowed to view the reviewers. Only
            a chair or a reviewer in this track can do that
        "404":
          description: Not Found. The specified paper or user does not exist.
...
  /papers/by-reviewer:
    get:
      tags:
      - Reviews
      - Assignments
      summary: Gets all papers a reviewer (the requester) is assigned to.
      description: Responds with a list of papers a user is assigned to in all tracks.
        This endpoint can be called by any reviewer.
      operationId: getAssignedPapers
      parameters:
      - name: requesterID
        in: query
        description: The ID of the user making the request.
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: Success.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PaperSummaryWithID'
        "500":
          description: Internal Server Error. An unexpected server error occurred.
        "404":
          description: Not Found. Such user does not exist
...
components:
  schemas:
...
    PaperSummaryWithID:
      type: object
      properties:
        paperID:
          type: integer
          description: The ID of the paper
          format: int64
          example: 2
        title:
          type: string
          description: The title of the paper
          example: Tails of frogs
        abstractSection:
          type: string
          description: The abstract of the paper
          example: Some sample abstract
      description: "Includes the title, abstract and ID of a paper"
...